working on the multiTouch assignment

+++ LayoutTesting/app/src/main/java/mobile/data/usage/spyspyyou/layouttesting/game/entities/Player.java
@@ -43,7 +43,7 @@ public class Player extends Entity {
         return ally;
     }
 
-    public int getRadius(){
+    public float getRadius(){
         return width / 2;
     }
